Initial Point of Code
Starting Point for project. Concept is to have a list of locations you
plan on going to. With these you add the start and end point with the
time you need to leave at. With google maps api you check both
locations for accuracy and then the time it would take to get from
point a to b. This will eventually be stored in a database to give you
the exact time you need to leave at to arrive in time.