Changed Direction of Concept
Site is now a 3 page website that does not require a user to log in.
Instead it just adds it to a database to be check along side all other
requests created. Requests are re-ran every ten minutes via a cron call
(that will be initiated soon). However, the site still checks to
locations but now also tells the user the time one would need to leave
as well as the duration of the trip. This will need to be