Events Update
Events page now has all fields validate for some type of input prior to
allowing the distance matrix to be ran. Once it runs it then reveals
the hidden two text areas (that can’t be changed) and enables the
submit button for the user to then enter the data to the server.
Next step is to enter the data into the server to be processed every 10
minutes.