Text Messaging and Directions
The site now allows the user to be texted right away of when he or she
should be leaving, as well as a url to give them directions.

The url is shortened using googles api for creating smaller url links.
Doing this to compensate for the amount of characters a text message
allows.

Last step will be to have a modal for success on both the contact and
submit page for user feedback.